+++ src/main/java/kitchenpos/dto/request/PutOrderTableEmptyRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.dto.request;
 
 public class PutOrderTableEmptyRequest {
