+++ src/main/java/kitchenpos/dto/response/MenuGroupResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.response;
+package kitchenpos.dto.response;
 
 public class MenuGroupResponse {
