+++ src/main/java/kitchenpos/dto/response/OrderLineItemDtoInOrderResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.response;
+package kitchenpos.dto.response;
 
 public class OrderLineItemDtoInOrderResponse {
