+++ src/main/java/kitchenpos/ui/MenuRestController.java
@@ -3,9 +3,9 @@ package kitchenpos.ui;
 import kitchenpos.application.MenuService;
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuGroup;
-import kitchenpos.ui.dto.CreateMenuRequest;
-import kitchenpos.ui.response.MenuGroupDto;
-import kitchenpos.ui.response.MenuResponse;
+import kitchenpos.dto.request.CreateMenuRequest;
+import kitchenpos.dto.response.MenuGroupDto;
+import kitchenpos.dto.response.MenuResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
