+++ src/main/java/kitchenpos/ui/ProductRestController.java
@@ -2,8 +2,8 @@ package kitchenpos.ui;
 
 import kitchenpos.application.ProductService;
 import kitchenpos.domain.Product;
-import kitchenpos.ui.dto.CreateProductRequest;
-import kitchenpos.ui.response.ProductResponse;
+import kitchenpos.dto.request.CreateProductRequest;
+import kitchenpos.dto.response.ProductResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
